[TASK] Unify directory listing methods
The two methods getFileList() and getFolderList() now rely on the same
internal mechanism to get their listing. This is roughly the same as in
FAL's local driver.

Currently, there is no caching of query results, which will most of the
time lead to double requests when listing a directory (once for both
files and folders). This could be tackled, but any operation changing
directory contents has to invalidate the cache then.