[FEATURE] Store passwords encrypted
The password for a WebDAV storage is currently stored unencrypted, which
might pose a security risk if it is exposed e.g. in a database dump.
Therefore, this commit moves the username and password to separate
fields and encrypts the password using Blowfish with Cyclic Block
Chaining (CBC).

Encryption uses a custom wrapper for the mcrypt library. This should be
integrated into the TYPO3 core. There is an existing issue for such a
library, see #21407.

The encrypted password is stored in a custom format including the used
method, encryption mode (CBC, EBC etc) and the initialization vector (if
any).

Resolves: #33303