[BUGFIX] Empty passwords throw warning
Empty passwords were encrypted by our helper methods, but lead to the
fourth part of the encrypted string (the actual password) being empty.
This made the decryption method struggle, which is now fixed by just
returning an empty string when encrypting an empty password (this case
is considered in the decryption routine already).

Additionally, this commit adds a unit test for the encryption class.