[BUGFIX] fileExists() returns true for folders
The WebDAV protocol does not differ hard between files and folders. The
only way to tell if a resource is a file or folder is by considering the
name - if it ends in a slash, it is most likely a folder, if it does
not, it is a file. This commit changes the behaviour of fileExists() to
check for a trailing slash; folderExists() already did this before.