[BUGFIX] SabreDAV exceptions are not caught
SabreDAV throws exceptions when errors occur (e.g. when a server does
not respond). These exceptions bubble up to the UI currently, leaving
the user with a broken backend and probably without any clue what to do.

This patch introduces exception handling in the WebDAV driver, though
currently a rather ugly one that just suppresses the exceptions and logs
a message.
Ideally, the exceptions would be transformed into more user-friendly
messages by the driver and then handed up to the UI to be displayed.

Change-Id: I951773f8a909b4ad620a3d5343293ec1a680a28d
Fixes: #41373