[BUGFIX] Missing bodies for PROPFIND requests
cURL does not seem to correctly set the options to retrieve PROPFIND
results after HEAD requests were sent within the same cURL context; see
<https://github.com/fruux/sabre-http/issues/47> for more information.

The bug was fixed in the meantime, but sabre/dav does still use an old
version of sabre/http. This commit fixes `composer.json` to include
the new version with an alias.