[TASK] Improve result caching
This adds caching of single file information from a PROPFIND request,
effectively making getFileInfo() 100% cached at least for the backend
file list. This should bring a huge speed improvement, as the number of
server requests will be dramatically lower.

Another step could be to request more than one level of folder
information, i.e. set a depth of 2 or three for DAV requests. This would
require more careful consideration for scenarios where parts of the
response are already in cache and other parts still need to be fetched.